stateMoveInChildComponent
