+++ src/components/Form.js
@@ -1,17 +1,41 @@
-import React from 'react'
-import SearchForm from './SearchForm'
+import React,{useState} from 'react'
+import KeywordForm from './KeywordForm'
 import Location from './Location'
 import SubmitButton from './Button'
 import SelectForm from './SelectForm'
+import { makeStyles } from '@material-ui/core/styles'
+
+const useStyles = makeStyles(() => ({
+  searchForm: {
+    // display: "flex",
+    textAlign: "center",
+  },
+}))
+
 const Form = (props) => {
+  const [searchKeyword, setSearchKeyword] = useState('')
+  const [searchLat, setSearchLat] = useState('')
+  const [searchLng, setSearchLng] = useState('')
+  const [searchGenre, setSearchGenre] = useState([])
+  const classes = useStyles()
+
+  const postWord = (e) => {
+    e.preventDefault()
+    props.postWord({
+      keyword: searchKeyword,
+      genre: searchGenre.code,
+      lat: searchLat,
+      lng: searchLng
+    })
+  }
   return (
-    <div>
-      <form onSubmit={props.postWord}>
-          <SelectForm genres={props.genres} setSearchGenre={props.setSearchGenre} searchGenre={props.searchGenre}/>
-            <SearchForm setSearchKeyword={props.setSearchKeyword} searchKeyword={props.searchKeyword} />
-            <SubmitButton />
-          <Location setSearchLat={props.setSearchLat} setSearchLng={props.setSearchLng} />
-          </form>
+    <div className={classes.searchForm}>
+      <form onSubmit={postWord}>
+        <SelectForm genres={props.genres} setSearchGenre={setSearchGenre} searchGenre={searchGenre}/>
+        <KeywordForm setSearchKeyword={setSearchKeyword} searchKeyword={searchKeyword} />
+        <SubmitButton />
+        <Location setSearchLat={setSearchLat} setSearchLng={setSearchLng} />
+      </form>
     </div>
   )
 }
