+++ src/components/KeywordForm.js
@@ -1,4 +1,4 @@
-import React, {useState} from 'react' 
+import React from 'react' 
 import TextField from '@material-ui/core/TextField'
 import { makeStyles } from '@material-ui/core/styles'
 
@@ -12,13 +12,11 @@ const useStyles = makeStyles((theme) => ({
   },
 }))
 
-const SearchForm = (props) => {
-  // const [keyword, setKeyword] = useState('')
+const KeywordForm = (props) => {
+  // console.log(props.searchKeyword)
   const classes = useStyles()
-  console.log(props.searchKeyword)
   // form中身の変更内容を取得
   const handleSearchKeywordChange = (e) => {
-    // setKeyword(e.target.value)
     props.setSearchKeyword(e.target.value)
   }
   return (
@@ -33,4 +31,4 @@ const SearchForm = (props) => {
   )
 }
 
-export default SearchForm
+export default KeywordForm
